requirements.txt
aiohttp
beautifulsoup4
pathvalidate
Pillow
requests
